Create BEAST2 XML Files from a Template

Use a template XML file to create new XML files for BEAST2 input. This function can replace sequences, sampling dates, mcmc iterations, frequency of storing trees, and, for a uniform clock rate, the minimum and maximum clock rates.

This was designed just to be used with the provided template file, though it could theoretically be used with others. The template file was written for tuberculosis data. It contains an adjustment for ascertainment bias due to the SNPs, a Gamma site model, and HKY substitution model, strict clock, Coalescent constant population model, uniform clock rate prior (min/max editable), and lognormal(1, 1.25) freqParameter, kappa, and popsize priors.

The template xml contains placeholders:

SNP_FILE_NAME_HERE for the name of the SNP file

ALIGNMENT_INFORMATION_HERE for the sequence information

DATE_INFORMATION_HERE for the sampling dates

CLOCKRATE_MINIMUM_HERE for the minimum uniform clock rate

CLOCKRATE_MAXIMUM_HERE for the maximum uniform clock rate

CLOCKRATE_INITIAL_HERE for the initial clock rate

MCMC_ITERATIONS_HERE for the number of MCMC iterations

STORE_EVERY_HERE for the frequency of storing trees

Usage

create_xml_files(
  path_to_template = system.file("xml_template/0_xml_template.xml", package =
    "beast2tpPipeline"),
  seqs_list,
  cluster_assignments,
  mcmc_iterations = 2e+07,
  store_every = 5000,
  min_clockrate = 10^(-8),
  max_clockrate = 5 * 10^(-7),
  init_clockrate = 10 * min_clockrate,
  whole_genome_length = 4.2 * 10^6,
  out_dir
)

Arguments

path_to_template

The path to the template XML file.

seqs_list

A list of matrices of DNA sequences (or SNPs), where rownames correspond to sample IDs.

cluster_assignments

A data frame of cluster assignments, including sample_id, cluster_name, and collectdt, as output by assign_snp_clusters. All sample IDs in the seqs matrices must be present in this data frame.

mcmc_iterations

The number of MCMC iterations to run.

store_every

The frequency of storing trees.

min_clockrate

The minimum clock rate for the uniform clock rate prior. Defaults to tuberculosis genome clock rate. DO NOT adjust for SNPs here.

max_clockrate

The maximum clock rate for the uniform clock rate prior. Defaults to tuberculosis genome clock rate.

init_clockrate

The initial clock rate for the clock rate.

whole_genome_length

The length of the whole genome. Used to adjust the clock rate for SNPs. Defaults to TB length.

out_dir

The directory to write the output XML file. Files are automatically named after their FASTA names.