Get Maximum Clade Credibility Trees Using Command-Line TreeAnnotator

Run TreeAnnotator to get the maximum clade credibility tree from a BEAST2 tree file.

Usage

get_mcctree(
  input_treesfile,
  output_dir,
  beast_iterations = 2e+07,
  burnin_fraction = 1/2,
  heights = "CA",
  treeannotator_path = "/Applications/\"BEAST 2.7.7\"/bin/treeannotator"
)

Arguments

input_treesfile

Character string. Path to the BEAST2 .trees file.

output_dir

Character string. Path to the desired output directory for the MCC tree nexus file.

beast_iterations

Numeric. Number of iterations in the BEAST2 run.

burnin_fraction

Numeric. Fraction of the MCMC chain to discard as burn-in.

heights

Character string. Node heights for the tree. Beware that if using anything other than common ancestor, could get some weird trees that won't work with TransPhylo.

treeannotator_path

Character string. Path to the TreeAnnotator executable.