Run TransPhylo on a Set of Trees

This is a wrapper for TransPhylo::infer_multittree_share_param. Can use it to run TransPhylo on a set of trees, either all MCC trees with parameter sharing, or a sample of posterior BEAST2 trees without parameter sharing (i.e., a sample of BEAST2 posterior trees for a single cluster, to incorporate some phylogenetic uncertainty).

Usage

run_TransPhylo(
  trees,
  cluster_name = NULL,
  type = c("mcctrees", "trees_sample"),
  cluster_dict,
  out_dir = "TransPhylo",
  output_name = "tp_res",
  gentime_shape = 10,
  gentime_scale = 1/10,
  sampling_shape = 10,
  sampling_scale = 1/10,
  prior_sampfrac_a = 1,
  prior_sampfrac_b = 19,
  start_off_p = 0.5,
  start_neg = 1.48,
  start_pi = 0.5,
  start_off_r = 1,
  share = c("neg", "off.r"),
  update_neg = TRUE,
  update_off_r = TRUE,
  update_off_p = FALSE,
  update_pi = TRUE,
  optiStart = 2,
  verbose = FALSE,
  delta_t = 0.01,
  mcmc_iterations = 1e+05,
  thinning = 10
)

Arguments

trees

An ape multiPhylo object containing the trees to be analyzed. If type is "mcctrees", this should be all MCC trees for all clusters, and the names of the trees need to match the names of the trees in the cluster dictionary. If type is "trees_sample", this should be a sample of posterior BEAST2 trees for a single cluster, and the cluster name should be provided as an argument.

cluster_name

If type is "trees_sample", this should be the name of the cluster for which the trees are being analyzed. This should match the cluster name in the cluster dictionary. If type is "mcctrees", this argument is ignored.

type

A character string specifying the type of trees to be analyzed. Can be "mcctrees" or "trees_sample". If "mcctrees", this function will run TransPhylo on all MCC trees for all clusters. If "trees_sample", this function will run TransPhylo on a sample of posterior BEAST2 trees for a single cluster.

cluster_dict

A data frame containing sample_id, cluster_name, and collectdt for each sequence, as output by assign_snp_clusters.

out_dir

A character string specifying the directory where the TransPhylo results will be saved.

output_name

A character string specifying the name of the output file.

gentime_shape

A numeric specifying the shape parameter of the gamma distribution for the generation time.

gentime_scale

A numeric specifying the scale parameter of the gamma distribution for the generation time.

sampling_shape

A numeric specifying the shape parameter of the gamma distribution for the sampling time. Default is the same as gentime_shape.

sampling_scale

A numeric specifying the scale parameter of the gamma distribution for the sampling time. Default is the same as gentime_scale.

prior_sampfrac_a

A numeric specifying the first shape parameter of the beta distribution for the sampling fraction.

prior_sampfrac_b

A numeric specifying the second shape parameter of the beta distribution for the sampling fraction.

start_off_p

A numeric specifying the starting value for the off.p parameter.

start_neg

A numeric specifying the starting value for the neg parameter.

start_pi

A numeric specifying the starting value for the pi parameter.

start_off_r

A numeric specifying the starting value for the off.r parameter.

share

A character vector specifying which parameters to share across clusters. Defaults to "neg" and "off.r". If type is "trees_sample", this parameter is ignored.

update_neg

A logical specifying whether to update the neg parameter.

update_off_r

A logical specifying whether to update the off.r parameter.

update_off_p

A logical specifying whether to update the off.p parameter.

update_pi

A logical specifying whether to update the pi parameter.

optiStart

A numeric specifying type of optimization to apply to MCMC start point.

verbose

A logical specifying whether to print verbose output.

delta_t

The grid size for the MCMC. Smaller is better but slower.

mcmc_iterations

A numeric specifying the number of MCMC iterations.

thinning

A numeric specifying the thinning parameter for the MCMC.

Value

A data frame containing the probability of each sample being an infection source.