Run TransPhylo on a Set of Trees

This is a wrapper for TransPhylo::infer_multittree_share_param. Can use it to run TransPhylo on a set of trees, either all MCC trees with parameter sharing, or a sample of posterior BEAST2 trees without parameter sharing (i.e., a sample of BEAST2 posterior trees for a single cluster, to incorporate some phylogenetic uncertainty).

Usage

run_TransPhylo(
  trees,
  cluster_name = NULL,
  type = c("mcctrees", "trees_sample"),
  cluster_dict,
  out_dir = "TransPhylo",
  output_name = "tp_res",
  ...
)

Arguments

trees

An ape multiPhylo object containing the trees to be analyzed. If type is "mcctrees", this should be all MCC trees for all clusters, and the names of the trees need to match the names of the trees in the cluster dictionary. If type is "trees_sample", this should be a sample of posterior BEAST2 trees for a single cluster, and the cluster name should be provided as an argument.

cluster_name

If type is "trees_sample", this should be the name of the cluster for which the trees are being analyzed. This should match the cluster name in the cluster dictionary. If type is "mcctrees", this argument is ignored.

type

A character string specifying the type of trees to be analyzed. Can be "mcctrees" or "trees_sample". If "mcctrees", this function will run TransPhylo on all MCC trees for all clusters. If "trees_sample", this function will run TransPhylo on a sample of posterior BEAST2 trees for a single cluster.

cluster_dict

A data frame containing sample_id, cluster_name, and collectdt for each sequence, as output by assign_snp_clusters.

out_dir

A character string specifying the directory where the TransPhylo results will be saved.

output_name

A character string specifying the name of the output file.

...

Additional arguments to be passed to TransPhylo::infer_multittree_share_param.

Value

A data frame containing the probability of each sample being an infection source.